Get user events or just the stream position

+++ BoxApi.V2/Model/Event.cs
@@ -17,7 +17,7 @@ namespace BoxApi.V2.Model
         /// <summary>
         /// The next position in the event stream that you should request in order to get the next events
         /// </summary>
-        public int NextStreamPosition { get; set; }
+        public long NextStreamPosition { get; set; }
 
         /// <summary>
         /// A collection of events that occured
